requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2025.4
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0
django-active-link==0.2.2
django-apscheduler==0.7.0
django-autoslug==1.9.9
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-crontab==0.7.1
django-debug-toolbar==5.1.0
django-import-export==4.3.7
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-qr-code==4.1.0
djangorestframework==3.16.0
docopt==0.6.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==5.3.2
mysql-connector-python==9.2.0
mysqlclient==2.2.4
num2words==0.5.14
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyMySQL==1.1.1
pypdf==5.4.0
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.1
querystring-parser==1.2.4
reportlab==4.4.0
requests==2.32.3
segno==1.6.6
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
